requirements.txt
requests~=2.31.0
cassandra-driver~=3.28.0